requirements.txt
blinker==1.7.0
chardet==5.2.0
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.2
gunicorn==21.2.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pillow==10.2.0
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2024.1
reportlab==4.1.0
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.4
Werkzeug==3.0.1
zipp==3.17.0
